Reorganised basically everything to facilitate this mod being a required dependency core mod.
