+++ common/src/main/java/com/collective/projectcore/items/CoreBaseItem.java
@@ -1,4 +1,4 @@
-package com.collective.items;
+package com.collective.projectcore.items;
 
 import net.minecraft.item.Item;
