+++ fabric/src/main/resources/fabric.mod.json
@@ -16,10 +16,10 @@
   "environment": "*",
   "entrypoints": {
     "main": [
-      "com.collective.fabric.ProjectCoreFabric"
+      "com.collective.projectcore.fabric.ProjectCoreFabric"
     ],
     "client": [
-      "com.collective.fabric.client.ProjectCoreFabricClient"
+      "com.collective.projectcore.client.ProjectCoreFabricClient"
     ]
   },
   "mixins": [
